refactor: move logging utilities to logging.py
